For #1828 - Move WebAppManifest to concept-engine
