+++ components/browser/session/src/main/java/mozilla/components/browser/session/engine/EngineObserver.kt
@@ -8,16 +8,15 @@ import android.graphics.Bitmap
 import android.os.Environment
 import mozilla.components.browser.session.Download
 import mozilla.components.browser.session.Session
-import mozilla.components.browser.session.manifest.WebAppManifestParser
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.HitResult
+import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.media.Media
 import mozilla.components.concept.engine.media.RecordingDevice
 import mozilla.components.concept.engine.permission.PermissionRequest
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.window.WindowRequest
 import mozilla.components.support.base.observer.Consumable
-import org.json.JSONObject
 
 /**
  * [EngineSession.Observer] implementation responsible to update the state of a [Session] from the events coming out of
@@ -155,13 +154,8 @@ internal class EngineObserver(
         media.unregisterObservers()
     }
 
-    override fun onWebAppManifestLoaded(manifest: JSONObject) {
-        val manifestResult = WebAppManifestParser().parse(manifest)
-        session.webAppManifest = if (manifestResult is WebAppManifestParser.Result.Success) {
-            manifestResult.manifest
-        } else {
-            null
-        }
+    override fun onWebAppManifestLoaded(manifest: WebAppManifest) {
+        session.webAppManifest = manifest
     }
 
     override fun onCrashStateChange(crashed: Boolean) {
