+++ components/concept/engine/build.gradle
@@ -26,6 +26,8 @@ android {
 }
 
 dependencies {
+    implementation project(':support-ktx')
+
     implementation Dependencies.kotlin_stdlib
     implementation Dependencies.androidx_annotation
     implementation Dependencies.kotlin_coroutines
