+++ components/concept/engine/src/test/resources/manifests/example_mdn.json
