+++ components/concept/engine/src/test/resources/manifests/minimal.json
