+++ components/concept/engine/src/test/resources/manifests/spec_typical.json
