+++ components/concept/engine/src/test/resources/manifests/unusual.json
