+++ components/feature/pwa/src/main/java/mozilla/components/feature/pwa/WebAppLauncherActivity.kt
@@ -10,7 +10,7 @@ import android.os.Bundle
 import androidx.annotation.VisibleForTesting
 import androidx.annotation.VisibleForTesting.PRIVATE
 import androidx.appcompat.app.AppCompatActivity
-import mozilla.components.browser.session.manifest.WebAppManifest
+import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.support.base.log.logger.Logger
 import mozilla.components.support.ktx.kotlin.toUri
