+++ components/feature/pwa/src/test/java/mozilla/components/feature/pwa/ext/ActivityKtTest.kt
@@ -6,7 +6,7 @@ package mozilla.components.feature.pwa.ext
 
 import android.app.Activity
 import android.content.pm.ActivityInfo
-import mozilla.components.browser.session.manifest.WebAppManifest
+import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.support.test.mock
 import org.junit.Test
 import org.mockito.Mockito.anyInt
