+++ docs/changelog.md
@@ -17,9 +17,11 @@ permalink: /changelog/
 
 * **browser-session**
   * Added handler for `onWebAppManifestLoaded` to update `session.webAppManifest`.
+  * Moved `WebAppManifest` to concept-engine.
 
 * **concept-engine**
-  * Added `onWebAppManifestLoaded` to `EngineSession`, called when the engine find a web app manifest.
+  * Added `onWebAppManifestLoaded` to `EngineSession`, called when the engine finds a web app manifest.
+  * Added `WebAppManifest` from browser-session.
 
 * **concept-sync**, **service-accounts**
   * ⚠️ **This is a breaking behavior change**: API changes to facilitate error handling; new method on AccountObserver interface.
